[Feat] 중복 숫자, 로또 번호 개수 초과 Error 추가
