+++ src/main/kotlin/lotto/exception/DuplicateInputException.kt
@@ -1,0 +1,3 @@
+package lotto.exception
+
+class DuplicateInputException : Exception()
