+++ src/main/kotlin/lotto/exception/InputCountException.kt
@@ -1,0 +1,3 @@
+package lotto.exception
+
+class InputCountException : Exception()
