+++ src/main/kotlin/lotto/util/InputUtil.kt
@@ -1,6 +1,8 @@
 package lotto.util
 
+import lotto.constant.LOTTO_NUM_COUNT
 import lotto.exception.DivisibleMoneyException
+import lotto.exception.InputCountException
 
 val br = System.`in`.bufferedReader()
 
@@ -12,34 +14,23 @@ fun readPaidMoney(): Int {
         }
         paidMoney
     } catch (e: Exception) {
-        mappingException(e)
+        showError(e)
         throw IllegalArgumentException()
     }
 }
 
 fun readWinLottoNumbers(): List<Int> {
     return try {
-        val winLottoNumbers = br.readLine().split(',').map { it.toInt() }
+        val winLottoNumbers = br.readLine().split(',').map { it.toInt() }.also { winLottoNumbers ->
+            if (winLottoNumbers.size != LOTTO_NUM_COUNT) throw InputCountException()
+        }
         winLottoNumbers
-    }catch (e: Exception){
-        mappingException(e)
+    } catch (e: Exception) {
+        showError(e)
         throw IllegalArgumentException()
     }
 }
 
-private fun mappingException(e: Exception){
-    when (e) {
-        is DivisibleMoneyException -> {
-            printErrorMessage(ErrorType.CANNOT_DIVISIBLE_MONEY)
-        }
-        is NumberFormatException -> {
-            printErrorMessage(ErrorType.WRONG_TYPE_INPUT)
-        }
-        else -> {
-            printErrorMessage(ErrorType.WRONG_VALUE)
-        }
-    }
-}
 
 fun bufferedReaderClose() {
     br.close()
